modularized function and docker logic
