+++ docker/.dockerignore
@@ -1,0 +1,3 @@
+
+node_modules
+npm-debug.log
