+++ docker/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "docker-test",
+  "version": "1.0.0",
+  "main": "server.js",
+  "scripts": {
+    "start": "node server.js"
+  },
+  "dependencies": {
+    "bluebird": "^3.4.6",
+    "body-parser": "^1.15.0",
+    "morgan": "^1.7.0",
+    "express": "^4.13.3",
+    "pg": "^4.5.3",
+    "pg-hstore": "^2.3.2",
+    "sequelize": "^3.21.0",
+    "retry": "^0.7.0"
+  }
+}
